monitor: Tach input to double
It was found that after the Value property on the Sensor.Value interface
was changed to be of double type, the tach input failed to get converted
correctly from double to int64 type when the property changed. Need to
set the tach input to be of double type explicitly.

Tested:
    Tach input values correctly read from dbus signal

+++ monitor/tach_sensor.hpp
@@ -94,7 +94,7 @@ class TachSensor
     /**
      * @brief Returns the input speed value
      */
-    inline int64_t getInput() const
+    inline double getInput() const
     {
         return _tachInput;
     }
@@ -274,7 +274,7 @@ class TachSensor
     /**
      * @brief The input speed, from the Value dbus property
      */
-    int64_t _tachInput = 0;
+    double _tachInput = 0;
 
     /**
      * @brief The current target speed, from the Target dbus property
